GpuDedisperser/pf-variance: derive pf-test weights from analytic PfAvarApproximation
Switch GpuDedisperser::test_one() to generate peak-finding test weights from the
analytic per-(subband, dm, profile) variances (PfAvarApproximation.tree_variance)
instead of variance_upper_bound()-based per-subband variances, via _make_weights().

Supporting API changes:
- ReferencePeakFindingKernel::_make_weights(): 'variances' arg is now Array<double>
  (base_weights stay float, computed with rsqrtf()).
- PfAvarApproximation ctor now takes shared_ptr<DedispersionPlan> (pybind binding
  and callers updated; GpuDedisperser passes the plan shared_ptr directly).
- test_one builds a PfAvarApproximation from the plan with unit freq_variances and
  feeds tree_variance[itree] to _make_weights(..., randomize=true); drops the old
  subband_variances setup. test_random re-enables float16 (force_float32 off).

+++ include/pirate/PeakFindingKernel.hpp
@@ -199,9 +199,9 @@ struct ReferencePeakFindingKernel
     //   base_weights[d,n,p] = 1/sqrt(variances[n,d,p]).
     // randomize=true scales each weight by a sparse random factor (see the .cu); randomize=false
     // uses the base_weights directly (no random multiplier).
-    //   out array shape       = (beams_per_batch, ndm_wt, nt_wt, nprofiles, fs.N)
-    //   variances array shape = (fs.N, ndm_wt, P)
-    void _make_weights(ksgpu::Array<float> &out, const ksgpu::Array<float> &variances, bool randomize);
+    //   out array shape       = (beams_per_batch, ndm_wt, nt_wt, nprofiles, fs.N)   (float)
+    //   variances array shape = (fs.N, ndm_wt, P)                                    (double)
+    void _make_weights(ksgpu::Array<float> &out, const ksgpu::Array<double> &variances, bool randomize);
 
     // At "level" l (where 0 <= l < log2(Wmax)), we have an array 'tmp_arr' containing input
     // array elements downsampled by 2^l (prepadded with data from the previous chunk).
